requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.3
bs4==0.0.1
click==7.1.2
Django==3.1.2
fpdf==1.7.2
gunicorn==20.0.4
importlib-metadata==2.0.0
joblib==0.17.0
Markdown==3.3.3
nltk==3.5
pytz==2020.1
PyYAML==5.3.1
regex==2020.10.28
soupsieve==2.0.1
sqlparse==0.4.1
tqdm==4.51.0
whitenoise==5.2.0
zipp==3.4.0
